+++ src/router/admin_observe.js
@@ -25,7 +25,7 @@ export const adminObserve = [
           {
             path: `viewer`,
             name: `${namePrefix}-log-viewer`,
-            component: () => import('@/views/observe/log/viewer/viewer'),
+            component: () => import('@/views/observe/log/viewer/index'),
             meta: {
               requireAuth: true,
               title: '日志查看器',
