+++ src/views/observe/log/viewer/components/LogEllipsisText.vue
