+++ src/views/resource/workload/components/WorkloadMonitor.vue
@@ -4,7 +4,7 @@
       <v-flex>
         <v-flex class="float-right">
           <v-sheet class="text-body-2 text--darken-1">
-            <BaseDatetimePicker2
+            <BaseDatetimePicker
               v-model="date"
               :default-value="30"
               @change="onDatetimeChange(undefined)"
