+++ src/views/tool/event/index.vue
@@ -4,7 +4,7 @@
       <template #extend>
         <v-flex class="kubegems__full-right align-center">
           <v-sheet class="text-body-2 text--darken-1">
-            <BaseDatetimePicker2
+            <BaseDatetimePicker
               v-model="date"
               :default-value="30"
               default-value-for-query
